[testloop] Refactor and rename some structs
