--- chain/client/src/test_utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod client;
-pub mod test_loop;
