+++ chain/network/src/lib.rs
@@ -28,7 +28,6 @@ pub mod shards_manager;
 pub mod state_sync;
 pub mod state_witness;
 pub mod tcp;
-pub mod test_loop;
 pub mod test_utils;
 pub mod types;
