+++ integration-tests/src/tests/client/challenges.rs
@@ -4,7 +4,7 @@ use near_chain::validate::validate_challenge;
 use near_chain::{Block, ChainStoreAccess, Error, Provenance};
 use near_chain_configs::Genesis;
 use near_chunks::shards_manager_actor::ShardsManagerActor;
-use near_client::test_utils::client::{create_chunk, create_chunk_with_transactions};
+use near_client::test_utils::{create_chunk, create_chunk_with_transactions};
 use near_client::{Client, ProcessTxResponse, ProduceChunkResult};
 use near_crypto::InMemorySigner;
 use near_network::types::NetworkRequests;
