+++ integration-tests/src/tests/client/resharding_v2.rs
@@ -3,7 +3,7 @@ use near_chain::Provenance;
 use near_chain::test_utils::wait_for_all_blocks_in_processing;
 use near_chain_configs::{Genesis, NEAR_BASE};
 use near_client::ProcessTxResponse;
-use near_client::test_utils::client::run_catchup;
+use near_client::test_utils::run_catchup;
 use near_crypto::InMemorySigner;
 use near_o11y::testonly::init_test_logger;
 use near_primitives::account::id::AccountId;
