+++ test-loop-tests/src/setup/mod.rs
@@ -1,4 +1,5 @@
 pub mod builder;
 pub mod drop_condition;
 pub mod env;
+pub mod setup;
 pub mod state;
