+++ test-loop-tests/src/setup/state.rs
@@ -12,10 +12,6 @@ use near_client::{PartialWitnessActor, ViewClientActorInner};
 use near_jsonrpc::ViewClientSenderForRpc;
 use near_network::shards_manager::ShardsManagerRequestFromNetwork;
 use near_network::state_witness::PartialWitnessSenderForNetwork;
-use near_network::test_loop::{
-    ClientSenderForTestLoopNetwork, TestLoopNetworkSharedState, TestLoopPeerManagerActor,
-    ViewClientSenderForTestLoopNetwork,
-};
 use near_network::types::{HighestHeightPeerInfo, PeerInfo};
 use near_parameters::RuntimeConfigStore;
 use near_primitives::block::GenesisId;
@@ -26,6 +22,11 @@ use near_primitives::upgrade_schedule::ProtocolUpgradeVotingSchedule;
 use near_store::Store;
 use nearcore::state_sync::StateSyncDumper;
 use tempfile::TempDir;
+
+use crate::utils::peer_manager_actor::{
+    ClientSenderForTestLoopNetwork, TestLoopNetworkSharedState, TestLoopPeerManagerActor,
+    ViewClientSenderForTestLoopNetwork,
+};
 
 use super::drop_condition::{DropCondition, TestLoopChunksStorage};
 
@@ -54,7 +55,7 @@ pub struct SharedState {
 
 /// This is the state associated with each node in the test loop environment before being built.
 /// The setup_client function will be called for each node to build the node and return TestData
-pub struct NodeState {
+pub struct NodeSetupState {
     pub account_id: AccountId,
     pub client_config: ClientConfig,
     pub store: Store,
@@ -65,7 +66,7 @@ pub struct NodeState {
 /// This state is specific to each node and is not shared across nodes.
 /// We can access each of the individual actors and senders from this state.
 #[derive(Clone)]
-pub struct TestData {
+pub struct NodeExecutionData {
     pub identifier: String,
     pub account_id: AccountId,
     pub peer_id: PeerId,
@@ -77,7 +78,7 @@ pub struct TestData {
     pub state_sync_dumper_handle: TestLoopDataHandle<StateSyncDumper>,
 }
 
-impl TestData {
+impl NodeExecutionData {
     pub fn get_highest_height_peer_info(
         &self,
         test_loop_data: &TestLoopData,
@@ -101,44 +102,44 @@ impl TestData {
     }
 }
 
-impl From<&TestData> for AccountId {
-    fn from(data: &TestData) -> AccountId {
+impl From<&NodeExecutionData> for AccountId {
+    fn from(data: &NodeExecutionData) -> AccountId {
         data.account_id.clone()
     }
 }
 
-impl From<&TestData> for PeerId {
-    fn from(data: &TestData) -> PeerId {
+impl From<&NodeExecutionData> for PeerId {
+    fn from(data: &NodeExecutionData) -> PeerId {
         data.peer_id.clone()
     }
 }
 
-impl From<&TestData> for ClientSenderForTestLoopNetwork {
-    fn from(data: &TestData) -> ClientSenderForTestLoopNetwork {
+impl From<&NodeExecutionData> for ClientSenderForTestLoopNetwork {
+    fn from(data: &NodeExecutionData) -> ClientSenderForTestLoopNetwork {
         data.client_sender.clone().with_delay(NETWORK_DELAY).into_multi_sender()
     }
 }
 
-impl From<&TestData> for ViewClientSenderForRpc {
-    fn from(data: &TestData) -> ViewClientSenderForRpc {
+impl From<&NodeExecutionData> for ViewClientSenderForRpc {
+    fn from(data: &NodeExecutionData) -> ViewClientSenderForRpc {
         data.view_client_sender.clone().with_delay(NETWORK_DELAY).into_multi_sender()
     }
 }
 
-impl From<&TestData> for ViewClientSenderForTestLoopNetwork {
-    fn from(data: &TestData) -> ViewClientSenderForTestLoopNetwork {
+impl From<&NodeExecutionData> for ViewClientSenderForTestLoopNetwork {
+    fn from(data: &NodeExecutionData) -> ViewClientSenderForTestLoopNetwork {
         data.view_client_sender.clone().with_delay(NETWORK_DELAY).into_multi_sender()
     }
 }
 
-impl From<&TestData> for PartialWitnessSenderForNetwork {
-    fn from(data: &TestData) -> PartialWitnessSenderForNetwork {
+impl From<&NodeExecutionData> for PartialWitnessSenderForNetwork {
+    fn from(data: &NodeExecutionData) -> PartialWitnessSenderForNetwork {
         data.partial_witness_sender.clone().with_delay(NETWORK_DELAY).into_multi_sender()
     }
 }
 
-impl From<&TestData> for Sender<ShardsManagerRequestFromNetwork> {
-    fn from(data: &TestData) -> Sender<ShardsManagerRequestFromNetwork> {
+impl From<&NodeExecutionData> for Sender<ShardsManagerRequestFromNetwork> {
+    fn from(data: &NodeExecutionData) -> Sender<ShardsManagerRequestFromNetwork> {
         data.shards_manager_sender.clone().with_delay(NETWORK_DELAY).into_sender()
     }
 }
