+++ test-loop-tests/src/tests/multinode_test_loop_example.rs
@@ -1,7 +1,6 @@
 use itertools::Itertools;
 use near_async::time::Duration;
 use near_chain_configs::test_genesis::{TestEpochConfigBuilder, ValidatorsSpec};
-use near_client::test_utils::test_loop::ClientQueries;
 use near_o11y::testonly::init_test_logger;
 use near_primitives::shard_layout::ShardLayout;
 use near_primitives::types::AccountId;
@@ -9,6 +8,7 @@ use near_primitives::types::AccountId;
 use crate::setup::builder::TestLoopBuilder;
 use crate::setup::env::TestLoopEnv;
 use crate::utils::ONE_NEAR;
+use crate::utils::client_queries::ClientQueries;
 use crate::utils::transactions::execute_money_transfers;
 
 const NUM_CLIENTS: usize = 4;
