+++ test-loop-tests/src/utils/receipts.rs
@@ -1,7 +1,7 @@
 use super::loop_action::LoopAction;
 use super::retrieve_client_actor;
 use super::sharding::{next_block_has_new_shard_layout, this_block_has_new_shard_layout};
-use crate::setup::state::TestData;
+use crate::setup::state::NodeExecutionData;
 use crate::utils::sharding::get_memtrie_for_shard;
 use near_async::test_loop::data::TestLoopData;
 use near_chain::ChainStoreAccess;
@@ -31,7 +31,7 @@ pub fn check_receipts_presence_at_resharding_block(
 ) -> LoopAction {
     let (checked_receipts, succeeded) = LoopAction::shared_success_flag();
     let action_fn = Box::new(
-        move |node_datas: &[TestData],
+        move |node_datas: &[NodeExecutionData],
               test_loop_data: &mut TestLoopData,
               client_account_id: AccountId| {
             let client_actor =
@@ -59,7 +59,7 @@ pub fn check_receipts_presence_after_resharding_block(
 ) -> LoopAction {
     let (checked_receipts, succeeded) = LoopAction::shared_success_flag();
     let action_fn = Box::new(
-        move |node_datas: &[TestData],
+        move |node_datas: &[NodeExecutionData],
               test_loop_data: &mut TestLoopData,
               client_account_id: AccountId| {
             let client_actor =
